+++ spec/active_interaction/concerns/active_modelable_spec.rb
@@ -1,5 +1,3 @@
-# coding: utf-8
-
 require 'spec_helper'
 
 shared_examples_for 'ActiveModel' do