+++ spec/active_interaction/filter_spec.rb
@@ -1,5 +1,3 @@
-# coding: utf-8
-
 require 'spec_helper'
 
 describe ActiveInteraction::Filter, :filter do